Reorganize qobj.validation for json schemas (#1983)
* Reorganize qobj.validation for json schemas

Reorganize the existing code for validating against .json schemas, which
was at `qiskit.qobj`, by:
* moving `_schema_validation.py` to under `qiskit.validation`, as the
  module was generic to all .json schemas.
* keep `validate_qobj_against_schema` under `qiskit.qobj`, since it is
  qobj-specific. The function is being used by most providers during
  their jobs.
* update tests and cleanup related machinery (exceptions, imports).

* Style fixes after merging
